# feat: Task 7.13 - Performance Tuning ve Caching Strategy