… .taskmaster/tasks/tasks.json
… apps/api/app/core/cache.py
… apps/api/tests/test_cache_performance.py